Refactor the unit test after refactoring the shared non-coin functions
